[REBRAND-SDK] Add MessageStream class in C#
