+++ Plugins/Marigold/Marigold.cs
@@ -4,7 +4,6 @@ using System.Runtime.InteropServices;
 using System.Collections.Generic;
 using System.Linq;
 using System;
-using MarigoldSimpleJSON;
 
 namespace MarigoldSDK {
 	public class Marigold : MonoBehaviour
@@ -34,27 +33,6 @@ namespace MarigoldSDK {
 
 		[DllImport("__Internal")]
 		private static extern void _setGeoIpTrackingDefault (bool enabled);
-		
-		[DllImport("__Internal")]
-		private static extern void _unreadCount ();
-
-		[DllImport("__Internal")]
-		private static extern void _messages ();
-
-		[DllImport("__Internal")]
-		private static extern void _showMessageDetail (string messageJSON);
-
-		[DllImport("__Internal")]
-		private static extern void _dismissMessageDetail ();
-
-		[DllImport("__Internal")]
-		private static extern void _registerImpression(string messageJSON, int impressionType);
-
-		[DllImport("__Internal")]
-		private static extern void _removeMessage (string messageJSON);
-
-		[DllImport("__Internal")]
-		private static extern void _markMessageAsRead (string messageJSON);
 
 		#endif
 		#endregion
@@ -156,92 +134,6 @@ namespace MarigoldSDK {
 			CallAndroidStatic("setGeoIpTrackingEnabled", enabled);
 			#endif
 		}
-		
-		/// <summary>
-		/// Asyncronously gets the unread count for the message stream.
-		/// Unread Count will be returned with an OnUnreadCountReceivedEvent - add a handler to handle this. 
-		/// Errors will be called back with an OnErrorEvent - add a handler to handle this. 
-		/// </summary>
-		public static void UnreadCount () {
-			#if UNITY_IOS
-			Marigold._unreadCount ();
-			#elif UNITY_ANDROID
-			CallAndroidStatic("unreadCount");
-			#endif
-		}
-		
-		/// <summary>
-		/// Asyncronously returns a list of Messages.
-		/// Messages will be returned with the OnMessagesReceivedEvent - add a handler to handle this. 
-		/// </summary>
-		public static void GetMessages() {
-			#if UNITY_IOS
-			Marigold._messages ();
-			#elif UNITY_ANDROID
-			CallAndroidStatic("getMessages");
-			#endif
-		}
-
-		/// <summary>
-		/// Shows the message detail for a given message
-		/// </summary>
-		/// <param name="message">The message for which you want to see the detail for</param>
-		public static void ShowMessageDetail (Message message) {
-			#if UNITY_IOS
-			Marigold._showMessageDetail (GetJsonForMessage(message).ToString());
-			#elif UNITY_ANDROID
-			CallAndroidStatic("showMessageDetail", message.messageID);
-			#endif
-		}
-
-		/// <summary>
-		/// Dismisses the message detail. Does nothing on Android.
-		/// </summary>
-		public static void DismissMessageDetail () {
-			#if UNITY_IOS
-			Marigold._dismissMessageDetail();
-			#elif UNITY_ANDROID
-			//Do Nothing
-			#endif
-		}
-
-		/// <summary>
-		/// Registers an impression for a given message.
-		/// </summary>
-		/// <param name="message">Marigold Message to create the impression on.</param>
-		/// <param name="type">The Type of impression to create.</param>
-		public static void RegisterImpression(Message message, ImpressionType type) {
-			#if UNITY_IOS
-			Marigold._registerImpression(GetJsonForMessage(message).ToString(), (int)type);
-			#elif UNITY_ANDROID
-			CallAndroidStatic("registerMessageImpression", GetJsonForMessage(message).ToString(), (int)type);
-			#endif
-		}
-
-		/// <summary>
-		/// Removes a given message from the user's stream or from a call to GetMessages.
-		/// Errors will be called back with an OnErrorEvent - add a handler to handle thi
-		/// </summary>
-		/// <param name="message">Message the message to be removed.</param>
-		public static void RemoveMessage(Message message) {
-			#if UNITY_IOS
-			Marigold._removeMessage(GetJsonForMessage(message).ToString());
-			#elif UNITY_ANDROID
-			CallAndroidStatic("removeMessage", GetJsonForMessage(message).ToString());
-			#endif
-		}
-
-		/// <summary>
-		/// Marks a given message as read.
-		/// Errors will be called back with an OnErrorEvent - add a handler to handle thi
-		/// </summary>
-		public static void MarkMessageAsRead (Message message) {
-			#if UNITY_IOS
-			Marigold._markMessageAsRead(GetJsonForMessage(message).ToString());
-			#elif UNITY_ANDROID
-			CallAndroidStatic("markMessageAsRead", GetJsonForMessage(message).ToString());
-			#endif
-		}
 
 		private static void CallAndroidStatic(string methodName, params object[] args) {
 			using (AndroidJavaClass androidWrapper = new AndroidJavaClass(AndroidMarigoldWrapper))
@@ -251,36 +143,7 @@ namespace MarigoldSDK {
 			}
 		}
 
-		/// <summary>
-		/// Used only by the underlying unity plugin code. Do not call.
-		/// </summary>
-		public void ReceiveError(string errorDescription) {
-			ErrorEventArgs args = new ErrorEventArgs ();
-			args.ErrorDescription = errorDescription;
-			OnErrorEvent.Invoke(this, args);
-		}
-
-		public void ReceiveMessagesJSONData(string messagesJSON) {
-			List<Message> messages = new List<Message>();
-			var jsonMessagesArray = JSON.Parse(messagesJSON);
-
-			for (int i = 0; i < jsonMessagesArray.Count; i++ ) {
-				Message m = new Message();
-				m.title = jsonMessagesArray[i]["title"];
-				m.messageID = jsonMessagesArray[i]["id"];
-				m.text = jsonMessagesArray[i]["text"];
-				m.URL = jsonMessagesArray[i]["url"];
-				m.videoURL = jsonMessagesArray[i]["videoURL"];
-				m.imageURL = jsonMessagesArray[i]["imageURL"];
-				m.type = (Message.MessageType)jsonMessagesArray[i]["type"].AsInt;
-				m.createdAt = jsonMessagesArray[i]["created_at"];
-				messages.Add(m);
-			}
-
-			MessagesReceivedEventArgs args = new MessagesReceivedEventArgs ();
-			args.messages = messages;
-			OnMessagesReceivedEvent(this, args);
-		}
+		
 		#endregion
 
 		#region Helpers 
@@ -292,17 +155,13 @@ namespace MarigoldSDK {
 			return dtDateTime;
 		}
 
-		public static JSONClass GetJsonForMessage (Message message) {
-			JSONClass jsonObject = new JSONClass();
-			if (message.messageID != null) jsonObject["id"] = message.messageID;
-			if (message.title != null) jsonObject["title"] = message.title;
-			if (message.text != null) jsonObject["text"] = message.text;
-			if (message.createdAt != null) jsonObject["created_at"] = message.createdAt;
-			if (message.URL != null) jsonObject["url"] = message.URL;
-			if (message.videoURL != null) jsonObject["card_media_url"] = message.videoURL;
-			if (message.imageURL != null) jsonObject["card_image_url"] = message.imageURL;
-			jsonObject["notifications"] = new JSONArray();
-			return jsonObject;
+		/// <summary>
+		/// Used only by the underlying unity plugin code. Do not call.
+		/// </summary>
+		public void ReceiveError(string errorDescription) {
+			MarigoldErrorEventArgs args = new MarigoldErrorEventArgs ();
+			args.ErrorDescription = errorDescription;
+			OnErrorEvent.Invoke(this, args);
 		}
 
 		/// <summary>
@@ -314,69 +173,25 @@ namespace MarigoldSDK {
 			OnDeviceIdReceivedEvent.Invoke(this, args);
 		}
 		
-		/// <summary>
-		/// Used only by the underlying unity plugin code. Do not call.
-		/// </summary>
-		public void ReceiveUnreadCount(string unreadCount) {
-			UnreadCountReceivedEventArgs args = new UnreadCountReceivedEventArgs ();
-			args.UnreadCount = unreadCount;
-			OnUnreadCountReceivedEvent.Invoke(this, args);
-		}
-		
 		#endregion
 		
 		#region Callbacks
-		public static event EventHandler<ErrorEventArgs> OnErrorEvent;
-		public static event EventHandler<MessagesReceivedEventArgs> OnMessagesReceivedEvent;
+		public static event EventHandler<MarigoldErrorEventArgs> OnErrorEvent;
 		public static event EventHandler<DeviceIDReceivedEventArgs> OnDeviceIdReceivedEvent;
-		public static event EventHandler<UnreadCountReceivedEventArgs> OnUnreadCountReceivedEvent;
 		#endregion
 	}
 
 	/// <summary>
 	/// Marigold error event arguments.
 	/// </summary>
-	public class ErrorEventArgs :EventArgs {
+	public class MarigoldErrorEventArgs :EventArgs {
 		public string ErrorDescription { get; set; }
 	}
-
-
-	/// <summary>
-	/// Marigold messages received event.
-	/// </summary>
-	public class MessagesReceivedEventArgs :EventArgs {
-		public List<Message> messages { get; set; }
-	}
-
-	/// <summary>
-	/// Marigold message class.
-	/// </summary>
-	public class Message {
-		// TODO - compare with native values
-		public string messageID { get; set; }
-		public string title { get; set; }
-		public string text { get; set; }
-		public string URL { get; set; }
-		public string videoURL { get; set; }
-		public string imageURL { get; set; }
-		public string createdAt { get; set; }
-		public MessageType type { get; set; }
-		public enum MessageType {Text, Image, Link, Video, FakePhoneCall, Other};
-	}
-
-	public enum ImpressionType {InAppNotificationView, StreamView, DetailView};
 
 	/// <summary>
 	/// Marigold Device ID received event.
 	/// </summary>
 	public class DeviceIDReceivedEventArgs :EventArgs {
 		public string DeviceID { get; set; }
-	}
-
-	/// <summary>
-	/// Marigold unread count received event.
-	/// </summary>
-	public class UnreadCountReceivedEventArgs :EventArgs {
-		public string UnreadCount { get; set; }
 	}
 }
