+++ Plugins/Marigold/MessageStream.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1d13ec7f04b74fbaabee2fea02de852
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
